requirements.txt
# conda install pytorch==0.4.1 torchvision cuda90 -c pytorch
# conda install -c intel mkl_fft==1.0.15
# pip install -i https://pypi.doubanio.com/simple/ -r requirements.txt 
absl-py==0.9.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
alembic==1.7.4
apptools==5.1.0
astor==0.8.1
astroid==2.4.2
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
autobahn==22.6.1
Automat==20.2.0
autopage==0.4.0
Babel==2.9.1
backcall==0.1.0
backports.entry-points-selectable==1.1.0
bleach==3.1.1
cachetools==4.0.0
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.1.0
click==8.0.3
cliff==3.9.0
cmaes==0.8.2
# cmd2==2.2.0
colorama==0.4.4
colorlog==6.5.0
configobj==5.0.6
constantly==15.1.0
cryptography==37.0.4
cycler==0.11.0
DateTime==4.3
decorator==4.4.2
defusedxml==0.6.0
dgl==0.4.2
dijkstra3d==1.12.0
distlib==0.3.3
docutils==0.17.1
edt==2.1.1
emoji==1.6.1
entrypoints==0.3
envisage==6.0.1
filelock==3.3.1
flake8==4.0.1
fonttools==4.34.4
frozenlist==1.3.0
future==0.18.2
gast==0.2.2
gensim==3.8.3
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grave==0.0.3
greenlet==1.1.2
grpcio==1.27.2
h5py==2.10.0
huggingface-hub==0.0.19
hyperlink==21.0.0
identify==2.3.1
idna==3.3
imageio==2.10.1
imagesize==1.2.0
importlib-metadata==4.2.0
importlib-resources==5.2.2
incremental==21.3.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
jedi==0.16.0
Jinja2==2.11.1
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.4.4
lazy-object-proxy==1.4.3
littleutils==0.2.2
llvmlite==0.37.0
Mako==1.1.5
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.5.2
mayavi==4.7.2 # need vtk 
mccabe==0.6.1
mistune==0.8.4
# mkl-fft==1.0.15   conda install -c intel mkl_fft==1.0.15
# mkl-random==1.1.0
# mkl-service==2.3.0
mlab==1.1.4
multidict==6.0.2
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nodeenv==1.6.0
notebook==6.0.3
numba==0.54.1
numpy==1.19.1
oauthlib==3.1.0
# ogb==1.3.2
open3d-python==0.7.0.0
opencv-python==3.4.2.17
opt-einsum==3.1.0
optuna==2.4.0
outdated==0.2.1
packaging==21.3
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.4.0
plotly==4.5.4
pptk==0.1.0
pre-commit==2.15.0
prettytable==2.2.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyface==7.3.0
pyflakes==2.4.0
Pygments==2.10.0
pylint==2.6.0
pymia==0.2.3
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.15.7
python-dateutil==2.8.2
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.2
pyzmq==19.0.0
regex==2021.10.23
releases==1.6.3
# requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
sacremoses==0.0.46
scikit-image==0.18.3
scikit-learn==1.0
scipy==1.4.1
semantic-version==2.6.0
Send2Trash==1.5.0
sentencepiece==0.1.96
SimpleITK==1.2.4
six==1.16.0
sklearn==0.0
smart-open==5.2.1
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.26
stevedore==3.5.0
tabulate==0.8.9
tensorboard==1.14.0
tensorboardX==1.9
tensorflow==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.4
threadpoolctl==3.0.0
tifffile==2021.10.12
tokenizers==0.10.3
toml==0.10.2
# torch==0.4.1   conda install pytorch==0.4.1 torchvision cuda90 -c pytorch
torchvision==0.2.2
tornado==6.0.4
tqdm==4.19.9
traitlets==4.3.3
traits==6.2.0
traitsui==7.2.0
# transformers==4.11.3
Twisted==22.4.0
txaio==22.2.1
typed-ast==1.4.1
typing-extensions==4.3.0
urllib3==1.25.8
virtualenv==20.9.0
vtk==9.0.1
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
wslink==1.6.6
yarl==1.7.2
zipp==3.5.0
zope.interface==5.4.0